removing dangling images

+++ src/orm/cleanup.py
@@ -1,16 +1,22 @@
 import os
 
-import config
+import itertools
+
 from orm import TestImage, orm, ReferenceName, get_base_path
 
 
 @orm.db_session
 def clean_up_unused_images():
-    images_to_delete = (x for x in TestImage.select() if not x.tests)
     base_path = get_base_path()
-    for image in images_to_delete:
-        path = os.path.join(base_path, image.filename)
-        image.delete()
+    images_to_delete = ((x, x.filename) for x in TestImage.select() if not x.tests)
+
+    all_images = [x.filename for x in TestImage.select()]
+    dangling_images = ((None, file) for file in os.listdir(base_path) if file not in all_images)
+
+    for image, filename in itertools.chain(images_to_delete, dangling_images):
+        path = os.path.join(base_path, filename)
+        if image:
+            image.delete()
         print('removing {0}'.format(path))
         os.remove(path)
 
@@ -32,3 +38,4 @@ def fix_reference_names(names):
 
 if __name__ == '__main__':
     clean_up_unused_images()
+
